+++ src/FunFair.BuildCheck.ProjectChecks/Settings/LoggingExtensions/ProjectsShouldHaveTrimAnalyzerConfiguredPolicyLoggingExtensions.cs
@@ -10,4 +10,15 @@ internal static partial class ProjectsShouldHaveTrimAnalyzerConfiguredPolicyLogg
         string projectName,
         string property
     );
+
+    [LoggerMessage(
+        EventId = 2,
+        Level = LogLevel.Error,
+        Message = "{projectName}: {property} must be true when IsTrimmable is true."
+    )]
+    public static partial void ProjectMustEnableTrimAnalyzerWhenTrimmable(
+        this ILogger<ProjectsShouldHaveTrimAnalyzerConfiguredPolicy> logger,
+        string projectName,
+        string property
+    );
 }