+++ CHANGELOG.md
@@ -6,13 +6,14 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 -->
 
 ## [Unreleased]
+### Security
 ### Added
+- Added TestProjectsShouldImportUnitTestsProps check to ensure test projects include UnitTests.props import
 ### Fixed
 ### Changed
 - SDK - Updated DotNet SDK to 10.0.200
 ### Removed
 ### Deployment Changes
-
 <!--
 Releases that have at least been deployed to staging, BUT NOT necessarily released to live.  Changes should be moved from [Unreleased] into here as they are merged into the appropriate release branch
 -->