+++ CHANGELOG.md
@@ -6,13 +6,16 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 -->
 
 ## [Unreleased]
+### Security
 ### Added
+- Test projects must import the UnitTests.props file
 ### Fixed
+- Disabled local publish step for multi-target-framework tool project to fix pre-commit hook
 ### Changed
 - SDK - Updated DotNet SDK to 10.0.200
+### Deprecated
 ### Removed
 ### Deployment Changes
-
 <!--
 Releases that have at least been deployed to staging, BUT NOT necessarily released to live.  Changes should be moved from [Unreleased] into here as they are merged into the appropriate release branch
 -->